Print statements and output.
Added lots of prints to get a better idea of what's going on. 
It looks like alot of the FOE is programmed as if it's facing is constant or always in one direction when checking for the next move.
Current output is an infinite loop of:

Current position is (0,0)
Moving left seeking fontier cell.
Current position is (0,1)
Moving left seeking fontier cell.
Current position is (0,1)
Moving left seeking fontier cell.
Current position is (0,0)
Moving left seeking fontier cell.